pyproject.toml
[tool.poetry]
name = "rest-shop"
version = "0.1.0"
description = "REST API for study project"
authors = ["Lysak Tatiana <lysak.tatiana2501@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.7"
python-environ = "^0.4.54"
djangorestframework = "^3.12.2"
drf-yasg = "^1.20.0"
mysqlclient = "^2.0.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
flakehell = "^0.9.0"
black = "^20.8b1"
Faker = "^6.6.0"

[tool.black]
max_line_length = 120

[tool.flakehell]
exclude = ["README.rst", "README.md", '**/migrations/**']
format = "colored"
max_line_length = 120
show_source = true

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-darglint = ["+*"]
flake8-docstrings = ["+*"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*", "-W503", "-E203", "-E501"]
pyflakes = ["+*"]
pylint = ["+*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
